requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.2.0
dj-database-url==0.4.2
Django==2.0.1
django-debug-toolbar==1.9.1
django-dotenv==1.4.2
django-jet==1.0.7
djangorestframework==3.7.7
docopt==0.6.2
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
oauth2client==4.1.2
psycopg2==2.7.3.2
pycodestyle==2.3.1
python-decouple==3.1
pytz==2017.3
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
